+++ src/components/atom/input/Input.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import classNames from 'classnames';
 import css from './Input.module.scss';
+import PropTypes from 'prop-types';
 
-const Input = (props) => {
-  const {disabled, type} = props;
-  const isCheckOrRadio = type==='checkbox' || type==='radio';
+const Input = ({ id, name, type, placeholder, value, onChange, disabled }) => {
+  const isCheckOrRadio = type === 'checkbox' || type === 'radio';
 
   const classes = classNames({
     [css['c-input']]: true,
@@ -15,12 +15,34 @@ const Input = (props) => {
 
   return <input
     className={classes}
-    {...props}
+    id={id}
+    name={name}
+    type={type}
+    placeholder={placeholder}
+    value={value}
+    onChange={onChange}
+    disabled={disabled}
   />
 };
 
+Input.propTypes = {
+  id: PropTypes.string,
+  name: PropTypes.string,
+  onChange: PropTypes.func,
+  placeholder: PropTypes.string,
+  type: PropTypes.string,
+  value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+  disabled: PropTypes.bool
+};
+
 Input.defaultProps = {
-  type: 'text'
+  id: null,
+  name: null,
+  onChange: f => f,
+  placeholder: null,
+  type: 'text',
+  value: null,
+  disabled: null
 };
 
 export default Input;