+++ src/components/organism/form/Form.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Form = ({ onSubmit, children }) => (
+  <form onSubmit={onSubmit}>{children}</form>
+);
+
+Form.propTypes = {
+  children: PropTypes.oneOfType([
+    PropTypes.array,
+    PropTypes.object,
+    PropTypes.string
+  ]).isRequired,
+  onSubmit: PropTypes.func.isRequired
+};
+
+Form.defaultProps = {};
+
+export default Form;