+++ src/index.js
@@ -2,19 +2,28 @@
 import { Grid, Col, Row } from 'react-flexbox-grid';
 
 // UI
+// atoms
 import Button from './components/atom/button/Button';
-import Form from './components/atom/form/Form';
 import Input from './components/atom/input/Input';
 import Label from './components/atom/label/Label';
-import FormGroup from './components/molecule/formGroup/FormGroup';
+
+// molecules
+import FormActions from './components/molecule/form-actions/FormActions';
+import FormGroup from './components/molecule/form-group/FormGroup';
+import FormRow from './components/molecule/form-row/FormRow';
+
+// organisms
+import Form from './components/organism/form/Form';
 
 export {
   Button,
+  Col,
   Form,
+  FormActions,
+  FormGroup,
+  FormRow,
+  Grid,
   Input,
   Label,
-  FormGroup,
-  Row,
-  Col,
-  Grid
+  Row
 };